requirements.txt
discord.py
asyncio
aiohttp
pprint
configparser
bs4
psutil
mysql-connector